requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.24.0

# Web and file handling
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Optional dependencies (install if needed)
# websockets>=12.0
# sqlalchemy>=2.0.0
# alembic>=1.13.0
# redis>=5.0.0
# celery>=5.3.0
# numpy>=1.24.0
# pandas>=2.0.0
# scikit-learn>=1.3.0
# openai>=1.0.0
# anthropic>=0.7.0
